arm64: dts: allwinner: a64: Orange Pi Win: Fix USB
The Orange Pi Win has 5 USB ports:
Four are standard A sockets, connected to an on-board HUB. The hub's and
socket's power regulators are enabled by GPIO PD7.
Also it features a microB socket, connected to the SoC's USB-OTG port.
It's power is supplied by the AXP PMIC, and the ID pin is connected to
GPIO PH9. It can serve both as a host or a client port.

Signed-off-by: Samuel Holland <samuel@sholland.org>
Signed-off-by: Andre Przywara <andre.przywara@arm.com>