KVM: arm/arm64: Avoid lpi_list lookup on guest exit
If we emulate an ITS interrupt controller for the guest, we keep all the
associated LPI IRQs in a linked list. Typically this list is not too
long, but can reach several hundred MSIs if the guest is in some serious
I/O business.
We don't need to traverse the list too often, but we do so on every
guest exit if we had injected an LPI before. Not only does this take
quite a while if the list is quite long, but also it potentially trashes
a good part of the cache, while holding the per-VM lpi_list_lock.

In an I/O-heavy use case (passing through NVMe disks while running
memcached in the guest, for instance), this leads to some considerable
performance impact, also limits scalability for guests with many VCPUs.

To avoid the global LPI list lookup in this case, we can just iterate
through our per-VCPU ap_list, which must hold the LPI we are looking
for. This list is typically very short, also just needs a per-VCPU lock.
As we don't need to touch the lpi_list at all, we avoid all the
disadvantages mentioned above.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>