arm64: dts: allwinner: Add Allwinner H616 .dtsi file
This (relatively) new SoC is similar to the H6, but drops the (broken)
PCIe support and the USB 3.0 controller. It also get the management
controller removed, which in turn removes *some*, but not all of the
devices formelly dedicated to the ARISC (CPUS).
There does not seem to be an external interrupt controller anymore, so
no external interrupts through an NMI pin. The AXP driver needs to learn
living with that.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>