mmc: sunxi: Add support for Allwinner H616 variant
The MMC controller in the H616 SoC expects the addresses in the DMA
descriptors to address words, not bytes, apparently to cover memory
beyond 4GB.
Catch the new compatible string, and shift the DMA addresses by 2 bits
in this case.
We could extend the DMA mask, but leave it at 32-bits for now.

Signed-off-by: Jernej Skrabec <jernej.skrabec@siol.net>
Signed-off-by: Andre Przywara <andre.przywara@arm.com>