clk: sunxi-ng: h6: Fix clock divider range on some clocks
While comparing clocks between the H6 and H616, some of the M factor
ranges were found to be wrong: the manual says they are only covering
two bits [1:0], but our code had "5" in the number of bits field.

By writing 0xff into that register in U-Boot and via FEL, it could be
confirmed that bit [4:2] are indeed masked off, so the manual is right.

Change to number of bits in the affected clock's description.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>