ac200-phy: strip down into pure PHY driver
With the original RFC approach, there is a platform driver for the PHY,
which talks both via MDIO to configure some AC200 PHY pecularities, but
also via the I2C regmap to configure and enable the glue logic.

Strip off the latter part, to make this a pure PHY driver.
The platform glue logic part will be a separate driver.