net: phy: mdio: add clock support for PHYs
So far the generic Ethernet PHY subsystem supports PHYs having a reset
line, which needs to be de-asserted before the PHY can be used. This
corresponds to an "RST" pin on most external PHY chips.
But most PHY chips also need an external clock signal, which may feed
some internal PLL, and/or is used to drive the internal logic. In many
systems this clock signal is provided by a fixed crystal oscillator, so
is of no particular interest to software.

However some system use a more complex clock source, or try to save a
few pennies by avoiding the crystal. The X-Powers AC200 mixed signal PHY
chip for instance uses a software-controlled clock gate, and the Lindenis
V5 development board drives its RTL8211 PHY via a clock pin on the SoC.

On those systems the clock source needs to be actively enabled before
the PHY can be used. To support those systems, add a struct clk,
populate it from the devicetree (if provided), and enabled or disable
it when needed, similar to toggeling the reset lines.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>