clk: sunxi-ng: Add support for the A523/T527 CCU
Add the clocks of the main CCU of the Allwinner A523 and T527 SoCs.
The A523 clocks were modelled after the manual, and double checked by
writing all 1's into the respective register, to spot all implemented
bits. The T527 is a superset of the A523, and there does not seem to be a
separate T527 manual available. Information was gathered from the BSP,
by probing registers, and by looking at other recent SoCs.

The PLL and mod clocks for the two CPU clusters and the DSU are part of
a separate CCU, also most audio clocks are collected in a DSP CCU, so
both of these clock groups are missing from this driver.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>