mmc: sunxi-mmc: Fix A100 compatible description
It turns out that the Allwinner A100/A133 SoC only supports 8K DMA
blocks (13 bits wide), for both the SD/SDIO and eMMC instances.
And while this would make a trivial fix, the H616 falls back to the A100
compatible string, so we have to now match the H616 compatible string
explicitly against the description advertising 64K DMA blocks.

So change the A100 descripion to restrict DMA blocks to 8K, and
introduce a new H616 description, carrying over the 64K block size.
Also remove the redundant setting of clk_delays to NULL on the way.

Fixes: 3536b82e5853 ("mmc: sunxi: add support for A100 mmc controller")
Signed-off-by: Andre Przywara <andre.przywara@arm.com>